+++ scripts/bumpChangelogs/cli.js
@@ -1,10 +1,10 @@
 const { prompt } = require('enquirer');
-const initialPrompt = require('./initialPrompt');
-const stageFile = require('./stageFile');
-const commitChanges = require('./commitChanges');
-const getChangelogCommitMessage = require('./getChangelogCommitMessage');
-const updateChangelogs = require('./updateChangelogs');
-const promptStageAndCommit = require('./promptStageAndCommit');
+const initialPrompt = require('../utilities/initialPrompt');
+const stageFile = require('../utilities/stageFile');
+const commitChanges = require('../utilities/commitChanges');
+const getChangelogCommitMessage = require('../utilities/getChangelogCommitMessage');
+const bumpChangelogs = require('.');
+const promptStageAndCommit = require('../utilities/promptStageAndCommit');
 
 const promptChanges = async ({ packageNames }) => {
   if (!packageNames.length) throw new Error('No packages selected');
@@ -27,7 +27,7 @@ initialPrompt(
   "How would you like to enter which package's changelogs to update?",
 )
   .then(promptChanges)
-  .then(updateChangelogs)
+  .then(bumpChangelogs)
   .then(promptStageAndCommit)
   .then(({ packageNames, paths, shouldCommitChanges }) => {
     if (shouldCommitChanges) {
