+++ scripts/utilities/commitChanges.js
