+++ scripts/utilities/getNewChangelogContent.js
