+++ scripts/utilities/getPackagePath.js
