+++ scripts/utilities/initialPrompt.js
