+++ scripts/utilities/promptPackageInput.js
