+++ scripts/utilities/promptPackageMultiselect.js
