+++ scripts/utilities/promptStageAndCommit.js
