+++ scripts/utilities/readFile.js
@@ -1,0 +1,12 @@
+const fs = require('fs');
+
+module.exports = filePath =>
+  new Promise((resolve, reject) => {
+    fs.readFile(filePath, 'utf8', (error, content) => {
+      if (error) {
+        reject(error);
+      } else {
+        resolve(content);
+      }
+    });
+  });
