+++ scripts/utilities/regeneratePackageLocks.js
